HBD-9: Changed; |package| happybirthdates; Renamed.
