+++ app/src/main/java/com/example/android/happybirthdates/contactcreator/ContactCreatorViewModel.kt
@@ -14,15 +14,14 @@
  * limitations under the License.
  */
 
-package com.example.android.trackmysleepquality.contactcreator
+package com.example.android.happybirthdates.contactcreator
 
-import android.app.Application
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.android.trackmysleepquality.database.ContactDatabaseDao
-import com.example.android.trackmysleepquality.database.ContactPerson
+import com.example.android.happybirthdates.database.ContactDatabaseDao
+import com.example.android.happybirthdates.database.ContactPerson
 import kotlinx.coroutines.*
 import java.util.*
