+++ app/src/main/java/com/example/android/happybirthdates/contactdetails/ContactDetailsViewModel.kt
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package com.example.android.trackmysleepquality.contactdetails
+package com.example.android.happybirthdates.contactdetails
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.example.android.trackmysleepquality.database.ContactDatabaseDao
-import com.example.android.trackmysleepquality.database.ContactPerson
+import com.example.android.happybirthdates.database.ContactDatabaseDao
+import com.example.android.happybirthdates.database.ContactPerson
 
 /**
  * ContactDetailsFragment's ViewModel.
