+++ app/src/main/java/com/example/android/happybirthdates/database/ContactDatabase.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.example.android.trackmysleepquality.database
+package com.example.android.happybirthdates.database
 
 import android.content.Context
 import androidx.room.Database
