+++ app/src/main/res/layout/fragment_contact_creator.xml
@@ -25,7 +25,7 @@
     <data>
         <variable
             name="contactCreatorViewModel"
-            type="com.example.android.trackmysleepquality.contactcreator.ContactCreatorViewModel" />
+            type="com.example.android.happybirthdates.contactcreator.ContactCreatorViewModel" />
     </data>
 
     <!-- Start of the visible fragment layout using ConstraintLayout -->
